+++ pysimbotlib/__init__.py
@@ -1,2 +1,2 @@
 from . import Robot
-from . import Window
+from . import PySimbotMap
